LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A5_T3-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Cannot use the undeclared label 'LABEL_IN'.
S12.8_A5_T3

PASS Expected parsing failure

TEST COMPLETE
